feat: support range requests for metadata
