+++ crates/rattler_installs_packages/src/package_database.rs
@@ -6,6 +6,7 @@ use crate::{
     project_info::{ArtifactInfo, ProjectInfo},
     FileStore, NormalizedPackageName, Version,
 };
+use async_http_range_reader::AsyncHttpRangeReader;
 use elsa::sync::FrozenMap;
 use futures::{pin_mut, stream, StreamExt};
 use http::{header::CONTENT_TYPE, HeaderMap, HeaderValue, Method};
@@ -171,6 +172,39 @@ impl PackageDb {
                 return self.get_pep658_metadata::<A>(artifact_info).await;
             }
 
+            dbg!("trying to read: {}", &artifact_info.url);
+
+            // First check if we can lazily load the metadata
+            if A::supports_sparse_metadata() {
+                // Check if the artifact is the same type as the info.
+                let name = A::Name::try_as(&artifact_info.filename)
+                    .expect("the specified artifact does not refer to type requested to read");
+
+                dbg!("trying to lazy read: {}", &artifact_info.url);
+
+                match AsyncHttpRangeReader::new(
+                    self.http.client.clone(),
+                    artifact_info.url.clone(),
+                    16384,
+                )
+                .await
+                {
+                    Ok(mut reader) => match A::read_metadata_bytes(name, &mut reader).await {
+                        Ok((blob, metadata)) => {
+                            self.put_metadata_in_cache(&artifact_info, &blob)?;
+                            return Ok((artifact_info, metadata));
+                        }
+                        Err(err) => {
+                            tracing::warn!("failed to sparsely read wheel file: {err}, falling back to downloading the whole file");
+                        }
+                    },
+                    Err(e) => {
+                        dbg!("failed to open lazy stream", e);
+                    }
+                }
+            }
+
+            // Otherwise download the entire artifact
             let artifact = self
                 .get_artifact_with_cache::<A>(artifact_info, CacheMode::Default)
                 .await?;
@@ -334,7 +368,7 @@ async fn fetch_simple_api(http: &Http, url: Url) -> miette::Result<Option<Projec
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::artifact::Wheel;
+    use crate::wheel::Wheel;
     use crate::PackageName;
     use tempfile::TempDir;
